apps/backend/requirements.txt
# Core FastAPI and ASGI server
fastapi>=0.110.0
uvicorn[standard]>=0.27.0

# Database and ORM
sqlalchemy[asyncio]>=2.0.27
asyncpg>=0.29.0
alembic>=1.13.0

# Redis and caching
redis[hiredis]>=5.0.1

# Background processing
celery[redis]>=5.4.0

# Data validation and serialization
pydantic[email]>=2.8.0

# Authentication and security
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
bcrypt==4.1.2

# External API integrations
httpx>=0.27.0
aiohttp>=3.11.0
openai>=1.50.0
anthropic>=0.39.0

# Message Queue
aio-pika==9.3.1

# Security and Encryption
cryptography==41.0.7

# Data Processing
numpy>=1.26.0
pandas>=2.2.0

# Monitoring and logging
structlog==23.2.0

# Environment and configuration
python-dotenv==1.0.0

# Date and time utilities
python-dateutil==2.8.2

# Type checking and development utilities (in requirements-dev.txt)google-generativeai
